requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
ZODB==5.8.1
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.18.0
numpy==1.26.4